Add URL 2.3.1 to Cargo.toml and update the Proxy struct's timeout field to u64 in src/config.rs. Also, add a new 'proxy' table to Config.toml to allow users to configure a proxy

+++ src/config.rs
@@ -18,7 +18,7 @@ pub struct AppConfig {
 pub struct Proxy {
     pub switch: bool,
     pub proxy: String,
-    pub timeout: u16,
+    pub timeout: u64,
     pub retry: u8,
     pub ca_cert_file: Option<String>,
 }
